requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
python-binance
